+++ functions/cd.c
@@ -3,48 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   cd.c                                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tjehaes <tjehaes@student.42luxembourg      +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/13 08:36:11 by tjehaes           #+#    #+#             */
-/*   Updated: 2025/01/13 10:55:36 by tjehaes          ###   ########.fr       */
+/*   Updated: 2025/01/22 10:15:55 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
-
-/*void    execute_cd(char *inpt)
-{
-    int     i;
-    char    *path;
-    int     j;
-
-    i = 2;
-    j = 0;
-    g_env.exit_status = 0;
-    while (inpt[i] == ' ' || inpt[i] == '\t')
-        i++;
-    if (i == 2 && inpt[i++] != '\0')
-    {
-        printf("error: need a space between function and argument\n");
-        g_env.exit_status = 1;
-        return;
-    }
-    path = malloc(sizeof(char *) * ft_strlen(inpt));
-    while (inpt[i] && inpt[i] != ' ' && inpt[i] != '\t' && j < 255)
-        path[j++] = inpt[i++];
-    path[j] = '\0';
-    if (j == 0 || path == NULL)
-    {
-        ft_printf("error: cd needs a directory path\n");
-        g_env.exit_status = 1;
-        return;
-    }
-    if (chdir(path) != 0)
-    {
-        perror("cd");
-        g_env.exit_status = 1;
-    }
-}*/
 
 void	handle_cd_error(int error_type)
 {
