+++ functions/env.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tjehaes <tjehaes@student.42luxembourg      +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/13 08:45:19 by tjehaes           #+#    #+#             */
-/*   Updated: 2025/01/13 10:56:46 by tjehaes          ###   ########.fr       */
+/*   Updated: 2025/01/22 10:15:30 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,42 +70,3 @@ void	execute_env(char *command)
 			perror("waitpid");
 	}
 }
-
-/*void execute_env(char *command)
-{
-	pid_t pid;
-	int   i;
-
-	g_env.exit_status = 0;
-	i = 0;
-	while (command[i] >= 'a' && command[i] <= 'z' && i < 3)
-		i++;
-	if (command[i] >= 'a' && command[i] <= 'z')
-	{
-		printf("error: command does not exist\ndid you mean env ?\n");
-		g_env.exit_status = 127;
-		return ;
-	}
-	else if (command[i] != '\0')
-	{
-		printf("error: env dont take any argument\n");
-		g_env.exit_status = 1;
-		return ;
-	}
-	pid = fork();
-	if (pid == -1)
-	{
-		perror("fork");
-		return ;
-	}
-	else if (pid == 0)
-	{
-		print_environment_variables();
-		exit(0);
-	}
-	else
-	{
-		if (waitpid(pid, NULL, 0) == -1)
-			perror("waitpid");
-	}
-}*/
