+++ functions/unset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   unset.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tjehaes <tjehaes@student.42luxembourg      +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/13 09:12:15 by tjehaes           #+#    #+#             */
-/*   Updated: 2025/01/13 11:00:38 by tjehaes          ###   ########.fr       */
+/*   Updated: 2025/01/22 09:50:01 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,60 +35,6 @@ static void	extract_varname_u(char *input, char *varname, int *i)
 	varname[j] = '\0';
 }
 
-/*static void	remove_var_from_env(char *varname)
-{
-	unsigned long	j;
-	unsigned long	j;
-	unsigned long	len;
-	char			*env_var;
-	unsigned long	i;
-	int				varname_len;
-
-	j = 0;
-	while (environ[j] != NULL)
-	{
-		if (ft_strncmp(environ[j], varname, ft_strlen(varname)) == 0
-			&& environ[j][ft_strlen(varname)] == '=')
-		{
-			if (environ[j] != NULL && environ[j] != varname)
-				free(environ[j]);
-			while (environ[j + 1] != NULL)
-			{
-				environ[j] = environ[j + 1];
-				j++;
-			}
-			environ[j] = NULL;
-			return ;
-		}
-		j++;
-	}
-}*/
-/*static void	remove_var_from_env(char *varname)
-{
-	j = 0;
-	len = strlen(varname);
-	while (environ[j] != NULL)
-	{
-		env_var = environ[j];
-		if (strncasecmp(env_var, varname, len) == 0 && env_var[len] == '=')
-		{
-			// Vérifie si la variable a été allouée dynamiquement
-			if (env_var != varname)
-			{
-				// Libère la mémoire allouée pour la variable d'environnement
-				free(env_var);
-			}
-			while (environ[j + 1] != NULL)
-			{
-				environ[j] = environ[j + 1];
-				j++;
-			}
-			environ[j] = NULL;
-			return ;
-		}
-		j++;
-	}
-}*/
 static void	remove_var_from_env(char *varname)
 {
     int	i;
