+++ handle_command2.c
@@ -48,19 +48,6 @@ int	check_true_command(char *path, char **args)
 		return (0);
 }
 
-/*void	free_args(char *path, char **args)
-{
-	int	i;
-
-	i = 0;
-	free(path);
-	while (args[i])
-	{
-		free(args[i]);
-		i++;
-	}
-	free(args);
-}*/
 void free_args(char *path, char **args)
 {
     int i;
@@ -87,93 +74,6 @@ void free_args(char *path, char **args)
     }
 }
 
-
-void	exec_perror(char *str)
-{
-	perror(str);
-	g_env.exit_status = 1;
-}
-
-/*void	handle_external_command(char *command)
-{
-	char	**args;
-	char	*path;
-	pid_t	pid;
-	int		status;
-
-	args = ft_split(command, ' ');
-	if (!correct_command(args))
-	{
-		free_args(NULL, args);
-		return ;
-	}
-	path = command_path(args[0]);
-	if (check_true_command(path, args) == 1)
-		return ;
-	signal(SIGINT, handle_sigint);
-	pid = fork();
-	if (pid == 0)
-	{
-		signal(SIGINT, SIG_DFL);
-		execute_command(path, args);
-	}
-	else if (pid > 0)
-	{
-		waitpid(pid, &status, 0);
-		if (WIFEXITED(status) && WEXITSTATUS(status) != 0)
-			g_env.exit_status = 1;
-	}
-	else
-		exec_perror("fork");
-	free_args(path, args);
-}*/
-
-
-/*void	handle_external_command(char *command)
-{
-    char	**args;
-    char	*path;
-    pid_t	pid;
-    int		status;
-
-    args = ft_split(command, ' ');
-    if (!correct_command(args))
-    {
-        free_args(NULL, args);
-        return ;
-    }
-    path = command_path(args[0]);
-    if (!path || check_true_command(path, args) == 1)
-    {
-        free(path);
-        free_args(NULL, args);
-        return ;
-    }
-    signal(SIGINT, handle_sigint);
-    pid = fork();
-    if (pid == 0)
-    {
-        signal(SIGINT, SIG_DFL);
-        execute_command(path, args);
-    }
-    else if (pid > 0)
-    {
-        waitpid(pid, &status, 0);
-        if (WIFEXITED(status) && WEXITSTATUS(status) != 0)
-            g_env.exit_status = 1;
-    }
-    else
-        exec_perror("fork");
-    free(path);
-    free_args(NULL, args);
-}*/
-
-void update_exit_status(int status)
-{
-    if (WIFEXITED(status) && WEXITSTATUS(status) != 0)
-        g_env.exit_status = (WEXITSTATUS(status));
-}
-
 void execute_command_with_fork(char *path, char **args)
 {
     pid_t pid;
@@ -195,7 +95,6 @@ void execute_command_with_fork(char *path, char **args)
     else
         exec_perror("fork");
 }
-
 
 void handle_external_command(char *command)
 {
